requirements.txt
slackclient==2.2.0
wget>=3.2
flake8>=3.7.8